complete signing

+++ src/commands/init.rs
@@ -4,7 +4,7 @@ use super::Cli;
 use bitcoin::Network;
 use tracing::info;
 
-pub fn execute(cli: &Cli, port : u16, network: Network) {
+pub fn execute(cli: &Cli, port : u32, network: Network) {
     info!("init config to: {}", &cli.home);
     config::update_app_home(&cli.home);
     config::Config::default(port, network).save().unwrap();
